MVP-6028-3: remove source & refactor utils
