+++ packages/notifi-frontend-client/lib/client/index.ts
@@ -1,4 +1,3 @@
-export * from './ensureSource';
 export * from './ensureTarget';
 export * from './NotifiFrontendClient';
 export * from './clientFactory';
