+++ packages/notifi-frontend-client/lib/utils/crypto.ts
@@ -1,0 +1,10 @@
+export const normalizeHexString = (input: string): string => {
+  let result = input;
+  if (input !== '') {
+    result = input.toLowerCase();
+    if (!result.startsWith('0x')) {
+      result = '0x' + result;
+    }
+  }
+  return result;
+};
