+++ packages/notifi-frontend-client/lib/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './notNullOrEmpty';
-export * from './packFilterOptions';
 export * from './resolveRef';
 export * from './typeUtils';
+export * from './crypto';
+export * from './areIdsEqual';
