--- packages/notifi-frontend-client/lib/utils/packFilterOptions.ts
@@ -1,11 +1,0 @@
-import type { FilterOptions } from '../models';
-
-export const packFilterOptions = (
-  clientOptions: Readonly<FilterOptions> | null,
-): string => {
-  if (clientOptions === null) {
-    return '{}';
-  }
-
-  return JSON.stringify(clientOptions);
-};
